docs: Update from Unlicense to AGPL v3.0
The GNU Affero General Public License version 3 (AGPL v3.0) is a free software license published by the Free Software Foundation in November 2007. It is part of the GNU General Public License family and is designed specifically for software that is used over a network.

Key Features of AGPL v3.0:

Strong Copyleft: The AGPL is a strong copyleft license, meaning that any modified versions of the software must also be distributed under the same license. This ensures that the freedoms granted by the license are preserved in derivative works.

Network Use Requirement: A significant aspect of the AGPL is its requirement that if you modify the software and use it to provide a service (like a web application), you must make the complete source code of your modified version available to all users who interact with it. This is particularly relevant for cloud services and web applications, where users may not have direct access to the software being run.

Compatibility: The AGPL can be combined with other works governed by the GNU General Public License (GPL) version 3, but it has specific terms that apply to network software.

User Freedom: The license aims to protect user freedoms, ensuring that users can run, modify, and share the software freely.

Overall, the AGPL v3.0 is considered one of the most protective licenses for user freedom in the context of networked software.